fixed the observer pattern
