+++ src/frontend/ConsoleObserver.java
@@ -1,0 +1,8 @@
+package frontend;
+
+public class ConsoleObserver implements Observer {
+    @Override
+    public void update(String message) {
+        System.out.println("Notification: " + message);
+    }
+}
