+++ src/frontend/Observer.java
@@ -1,0 +1,5 @@
+package frontend;
+
+public interface Observer {
+    void update(String message);
+}
